+++ GRF/GrfCollection.cs
@@ -1,12 +1,12 @@
-﻿namespace GRF
-{
-    public class GrfCollection
-    {
-        public GrfCollection() { }
-        public GrfCollection( string iniFilePath ) => Load( iniFilePath );
+﻿//namespace GRF
+//{
+//    public class GrfCollection
+//    {
+//        public GrfCollection() { }
+//        public GrfCollection( string iniFilePath ) => Load( iniFilePath );
 
-        public void Load( string iniFilePath )
-        {
-        }
-    }
-}
+//        public void Load( string iniFilePath )
+//        {
+//        }
+//    }
+//}