+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## [0.1.1] - 2018-09-11
+## [Unreleased]
+## Changed
+- GrfFile.Name to FilePath
+
+### Added
+- GrfFile.FileName
+- GrfFile.FileType
+- Loading Version 0x102 GRF Files
+- Loading Version 0x103 GRF Files
+
+## 0.1.1 - 2018-09-11
 ### Added
 - Changelog
 - Readme with example usage
@@ -11,4 +21,5 @@
 - Loading Version 0x200 GRF Files
 - NuGet package https://www.nuget.org/packages/GRF/
 
+[Unreleased]: https://github.com/arminherling/GRF/releases/tag/v0.1.0...HEAD
 [0.1.0]: https://github.com/arminherling/GRF/releases/tag/v0.1.0